+++ packages/gov-ui-kit/src/utils/formatterUtils/index.ts
@@ -1,0 +1,2 @@
+export { formatterUtils, type IFormatNumberOptions } from './formatterUtils';
+export { NumberFormat, numberFormats, type INumberFormat } from './formatterUtilsDefinitions';